requirements.txt
opencv-python
pytesseract
pandas
numpy
scikit-learn
Pillow
